requirements.txt
pandas
plotly
pyyaml